+++ aldrin/src/lib.rs
@@ -90,7 +90,7 @@ pub use channel::{
 pub use client::Client;
 pub use discoverer::{Discoverer, DiscovererBuilder, DiscovererEvent, DiscovererEventKind};
 pub use error::Error;
-pub use handle::{Handle, PendingFunctionResult, PendingFunctionValue};
+pub use handle::Handle;
 pub use lifetime::{Lifetime, LifetimeId, LifetimeScope};
 pub use object::Object;
 pub use reply::Reply;